requirements.txt
discord.py
aiosqlite
easy_pil
aiogram
tweety-ns
python-dotenv
pyyaml